+++ packages/dara-components/changelog.md
@@ -2,7 +2,7 @@
 title: Changelog
 ---
 
-## NEXT
+## 1.16.23
 
 -  Fixed an issue with `CodeEditor`, `Input`, `Select`, `Textarea` components where they would bubble up Enter keypresses when confirming composition using IME input
 -  Fixed `UploadDropzone` error messages to properly distinguish between wrong file types and multiple files.